Change hash table of joins to InPlaceMutableHashTable
I wanted to change just the cached build side joins, but somehow it was
easier to change all the joins. We only have cached build side joins in
the paper, so it's fine for the experiments.

Note that the original hash table allowed for spilling, but this one
does not. But this is not an issue, because spilling doesn't make much
sense for a cached build side, since the assumption is that we probe
into it by random access, so even a few percent spilled would totally
kill the performance. And this hash table is faster than the original
because it doesn't have to deal with the possibility of spilling, so it
has a much simpler design.